move db to its own model
