+++ client/App.js
@@ -3,7 +3,7 @@ import React from 'react';
 import GameOverModal from './components/GameOverModal';
 import WelcomeModal from './components/WelcomeModal';
 
-import { firebase } from './firebase/index';
+// import { firebase } from './firebase/index';
 
 const PLAYER_RADIUS = 25;
 const JUNK_SIZE = 15;
@@ -277,11 +277,11 @@ export default class App extends React.Component {
     ctx.closePath();
 
     // Test Firebase connection
-    console.log(firebase);
-    console.log(firebase.database());
+    // console.log(firebase);
+    // console.log(firebase.database());
 
-    let ref = firebase.database().ref('leaderboard');
-    console.log(ref);
+    // let ref = firebase.database().ref('leaderboard');
+    // console.log(ref);
   }
 
   drawHoles() {
